+++ spiral/src/test/java/spirals/ulam/translators/DensityTranslatorTest.java
@@ -12,8 +12,8 @@ class DensityTranslatorTest {
     public void testDensityTranslation_1() {
         long [][] input = SimpleUlamGenerator.generateMatrix(3);
         boolean[][] booleans = BinaryTranslator.translateToBoolean(input);
-        long [][] expected = {{1, 3, 2}, {2, 4, 2}, {1, 2, 1}};
-        long [][] actual = DensityTranslator.translate(booleans);
+        short [][] expected = {{1, 3, 2}, {2, 4, 2}, {1, 2, 1}};
+        short [][] actual = DensityTranslator.translate(booleans);
         ConsolePrinter.prettyPrintMatrixWithFrames(input);
         assertArrayEquals(expected, actual);
     }
@@ -22,8 +22,8 @@ class DensityTranslatorTest {
     public void testDensityTranslation_2() {
         long [][] input = SimpleUlamGenerator.generateMatrix(5);
         boolean[][] booleans = BinaryTranslator.translateToBoolean(input);
-        long [][] expected = {{2, 2, 2, 2, 2}, {3, 3, 3, 4, 4}, {3, 3, 4, 3, 3}, {2, 3, 3, 3, 2}, {1, 2, 2, 1, 0}};
-        long [][] actual = DensityTranslator.translate(booleans);
+        short [][] expected = {{2, 2, 2, 2, 2}, {3, 3, 3, 4, 4}, {3, 3, 4, 3, 3}, {2, 3, 3, 3, 2}, {1, 2, 2, 1, 0}};
+        short [][] actual = DensityTranslator.translate(booleans);
         ConsolePrinter.prettyPrintMatrixWithFrames(input);
         assertArrayEquals(expected, actual);
     }
